More improvements and reuse of common elements in load/save
