+++ android/assets/jsons/translations/template.properties
@@ -768,6 +768,7 @@ Load from custom location =
 Save to custom location = 
 Could not save game to custom location! = 
 '[saveFileName]' copied to clipboard! = 
+Current game copied to clipboard! = 
 Could not save game to clipboard! = 
 Download missing mods = 
 Missing mods are downloaded successfully. =
