+++ core/src/com/unciv/ui/screens/savescreens/LoadGameScreen.kt
@@ -2,26 +2,19 @@ package com.unciv.ui.screens.savescreens
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.files.FileHandle
-import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton
-import com.badlogic.gdx.utils.GdxRuntimeException
-import com.badlogic.gdx.utils.SerializationException
 import com.unciv.Constants
-import com.unciv.UncivGame
 import com.unciv.logic.MissingModsException
 import com.unciv.logic.UncivShowableException
 import com.unciv.logic.files.PlatformSaverLoader
 import com.unciv.logic.files.UncivFiles
-import com.unciv.logic.github.Github
-import com.unciv.logic.github.Github.folderNameToRepoName
 import com.unciv.models.ruleset.RulesetCache
 import com.unciv.models.translations.tr
 import com.unciv.ui.components.UncivTooltip.Companion.addTooltip
 import com.unciv.ui.components.extensions.disable
 import com.unciv.ui.components.extensions.enable
 import com.unciv.ui.components.extensions.isEnabled
-import com.unciv.ui.components.extensions.toLabel
 import com.unciv.ui.components.extensions.toTextButton
 import com.unciv.ui.components.input.KeyCharAndCode
 import com.unciv.ui.components.input.keyShortcuts
@@ -33,89 +26,22 @@ import com.unciv.ui.popups.ToastPopup
 import com.unciv.utils.Concurrency
 import com.unciv.utils.Log
 import com.unciv.utils.launchOnGLThread
-import java.io.FileNotFoundException
-import java.nio.file.attribute.DosFileAttributes
-import kotlin.io.path.Path
-import kotlin.io.path.readAttributes
+import kotlinx.coroutines.CoroutineScope
 
 class LoadGameScreen : LoadOrSaveScreen() {
     private val copySavedGameToClipboardButton = getCopyExistingSaveToClipboardButton()
-    private val errorLabel = "".toLabel(Color.RED)
     private val loadMissingModsButton = getLoadMissingModsButton()
     private var missingModsToLoad: Iterable<String> = emptyList()
 
-    companion object {
+    /** Inheriting here again exposes [getLoadExceptionMessage] and [loadMissingMods] to
+     *  other clients (WorldScreen, QuickSave, Multiplayer) without needing to rewrite many imports 
+     */
+    companion object : Helpers {
         private const val loadGame = "Load game"
         private const val loadFromCustomLocation = "Load from custom location"
         private const val loadFromClipboard = "Load copied data"
         private const val copyExistingSaveToClipboard = "Copy saved game to clipboard"
         internal const val downloadMissingMods = "Download missing mods"
-
-        /** Gets a translated exception message to show to the user.
-         * @param file Optional, used for detection of local read-only files, only relevant for write operations on Windows desktops.
-         * @return The first returned value is the message, the second is signifying if the user can likely fix this problem. */
-        fun getLoadExceptionMessage(ex: Throwable, primaryText: String = "Could not load game!", file: FileHandle? = null): Pair<String, Boolean> {
-            val errorText = StringBuilder(primaryText.tr())
-            errorText.appendLine()
-            var cause = ex
-            while (cause.cause != null && cause is GdxRuntimeException) cause = cause.cause!!
-
-            fun FileHandle.isReadOnly(): Boolean {
-                try {
-                    val attr = Path(file().absolutePath).readAttributes<DosFileAttributes>()
-                    return attr.isReadOnly
-                } catch (_: Throwable) { return false }
-            }
-
-            val isUserFixable = when (cause) {
-                is UncivShowableException -> {
-                    errorText.append(ex.localizedMessage)
-                    true
-                }
-                is SerializationException -> {
-                    errorText.append("The file data seems to be corrupted.".tr())
-                    false
-                }
-                is FileNotFoundException -> {
-                    // This is thrown both for chmod/ACL denials and for writing to a file with the read-only attribute set
-                    // On Windows, `message` is already (and illegally) partially localized.
-                    val localizedMessage = UncivGame.Current.getSystemErrorMessage(5)
-                    val isPermissionDenied = cause.message?.run {
-                        contains("Permission denied") || (localizedMessage != null && contains(localizedMessage))
-                    } == true
-                    if (isPermissionDenied) {
-                        if (file != null && file.isReadOnly())
-                            errorText.append("The file is marked read-only.".tr())
-                        else
-                            errorText.append("You do not have sufficient permissions to access the file.".tr())
-                    }
-                    isPermissionDenied
-                }
-                else -> {
-                    errorText.append("Unhandled problem, [${ex::class.simpleName} ${ex.stackTraceToString()}]".tr())
-                    false
-                }
-            }
-            return Pair(errorText.toString(), isUserFixable)
-        }
-
-        fun loadMissingMods(missingMods: Iterable<String>, onModDownloaded:(String)->Unit, onCompleted:()->Unit) {
-            for (rawName in missingMods) {
-                val modName = rawName.folderNameToRepoName().lowercase()
-                val repos = Github.tryGetGithubReposWithTopic(10, 1, modName)
-                    ?: throw UncivShowableException("Could not download mod list.")
-                val repo = repos.items.firstOrNull { it.name.lowercase() == modName }
-                    ?: throw UncivShowableException("Could not find a mod named \"[$modName]\".")
-                val modFolder = Github.downloadAndExtract(
-                    repo,
-                    UncivGame.Current.files.getModsFolder()
-                )
-                    ?: throw Exception("Unexpected 404 error") // downloadAndExtract returns null for 404 errors and the like -> display something!
-                Github.rewriteModOptions(repo, modFolder)
-                onModDownloaded(repo.name)
-            }
-            onCompleted()
-        }
     }
 
     init {
@@ -152,7 +78,7 @@ class LoadGameScreen : LoadOrSaveScreen() {
     private fun Table.initRightSideTable() {
         add(getLoadFromClipboardButton()).row()
         addLoadFromCustomLocationButton()
-        add(errorLabel).width(stage.width / 2).row()
+        add(errorLabel).width(stage.width / 2).center().row()
         add(loadMissingModsButton).row()
         add(deleteSaveButton).row()
         add(copySavedGameToClipboardButton).row()
@@ -237,18 +163,7 @@ class LoadGameScreen : LoadOrSaveScreen() {
         copyButton.onActivation {
             val file = selectedSave ?: return@onActivation
             Concurrency.run(copyExistingSaveToClipboard) {
-                try {
-                    val gameText = file.readString()
-                    Gdx.app.clipboard.contents = if (gameText[0] == '{') Gzip.zip(gameText) else gameText
-                    launchOnGLThread {
-                        ToastPopup("'[${file.name()}]' copied to clipboard!", this@LoadGameScreen)
-                    }
-                } catch (ex: Throwable) {
-                    ex.printStackTrace()
-                    launchOnGLThread {
-                        ToastPopup("Could not save game to clipboard!", this@LoadGameScreen)
-                    }
-                }
+                copySaveToClipboard(file)
             }
         }
         copyButton.disable()
@@ -256,6 +171,31 @@ class LoadGameScreen : LoadOrSaveScreen() {
         copyButton.keyShortcuts.add(ctrlC)
         copyButton.addTooltip(ctrlC)
         return copyButton
+    }
+
+    private fun CoroutineScope.copySaveToClipboard(file: FileHandle) {
+        val gameText = try {
+            file.readString()
+        } catch (ex: Throwable) {
+            val (errorText, isUserFixable) = getLoadExceptionMessage(ex, saveToClipboardErrorMessage)
+            if (!isUserFixable)
+                Log.error(saveToClipboardErrorMessage, ex)
+            launchOnGLThread {
+                ToastPopup(errorText, this@LoadGameScreen)
+            }
+            return
+        }
+        try {
+            Gdx.app.clipboard.contents = if (gameText[0] == '{') Gzip.zip(gameText) else gameText
+            launchOnGLThread {
+                ToastPopup("'[${file.name()}]' copied to clipboard!", this@LoadGameScreen)
+            }
+        } catch (ex: Throwable) {
+            Log.error(saveToClipboardErrorMessage, ex)
+            launchOnGLThread {
+                ToastPopup(saveToClipboardErrorMessage, this@LoadGameScreen)
+            }
+        }
     }
 
     private fun getLoadMissingModsButton(): TextButton {
@@ -268,26 +208,20 @@ class LoadGameScreen : LoadOrSaveScreen() {
     }
 
     private fun handleLoadGameException(ex: Exception, primaryText: String = "Could not load game!") {
-        Log.error("Error while loading game", ex)
-        val (errorText, isUserFixable) = getLoadExceptionMessage(ex, primaryText)
+        val isUserFixable = handleException(ex, primaryText)
+        if (!isUserFixable) {
+            val cantLoadGamePopup = Popup(this@LoadGameScreen)
+            cantLoadGamePopup.addGoodSizedLabel("It looks like your saved game can't be loaded!").row()
+            cantLoadGamePopup.addGoodSizedLabel("If you could copy your game data (\"Copy saved game to clipboard\" - ").row()
+            cantLoadGamePopup.addGoodSizedLabel("  paste into an email to yairm210@hotmail.com)").row()
+            cantLoadGamePopup.addGoodSizedLabel("I could maybe help you figure out what went wrong, since this isn't supposed to happen!").row()
+            cantLoadGamePopup.addCloseButton()
+            cantLoadGamePopup.open()
+        }
 
-        Concurrency.runOnGLThread {
-            if (!isUserFixable) {
-                val cantLoadGamePopup = Popup(this@LoadGameScreen)
-                cantLoadGamePopup.addGoodSizedLabel("It looks like your saved game can't be loaded!").row()
-                cantLoadGamePopup.addGoodSizedLabel("If you could copy your game data (\"Copy saved game to clipboard\" - ").row()
-                cantLoadGamePopup.addGoodSizedLabel("  paste into an email to yairm210@hotmail.com)").row()
-                cantLoadGamePopup.addGoodSizedLabel("I could maybe help you figure out what went wrong, since this isn't supposed to happen!").row()
-                cantLoadGamePopup.addCloseButton()
-                cantLoadGamePopup.open()
-            }
-
-            errorLabel.setText(errorText)
-            errorLabel.isVisible = true
-            if (ex is MissingModsException) {
-                loadMissingModsButton.isVisible = true
-                missingModsToLoad = ex.missingMods
-            }
+        if (ex is MissingModsException) {
+            loadMissingModsButton.isVisible = true
+            missingModsToLoad = ex.missingMods
         }
     }
 
@@ -296,7 +230,7 @@ class LoadGameScreen : LoadOrSaveScreen() {
         descriptionLabel.setText(Constants.loading.tr())
         Concurrency.runOnNonDaemonThreadPool(downloadMissingMods) {
             try {
-                Companion.loadMissingMods(missingModsToLoad,
+                loadMissingMods(missingModsToLoad,
                     onModDownloaded = {
                         val labelText = descriptionLabel.text // Surprise - a StringBuilder
                         labelText.appendLine()
@@ -315,7 +249,9 @@ class LoadGameScreen : LoadOrSaveScreen() {
                     }
                 )
             } catch (ex: Exception) {
-                handleLoadGameException(ex, "Could not load the missing mods!")
+                launchOnGLThread {
+                    handleLoadGameException(ex, "Could not load the missing mods!")
+                }
             } finally {
                 launchOnGLThread {
                     loadMissingModsButton.isEnabled = true
