+++ core/src/com/unciv/ui/screens/savescreens/LoadOrSaveScreen.kt
@@ -1,11 +1,18 @@
 package com.unciv.ui.screens.savescreens
 
 import com.badlogic.gdx.files.FileHandle
+import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.scenes.scene2d.ui.CheckBox
 import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton
+import com.badlogic.gdx.utils.Align
+import com.badlogic.gdx.utils.GdxRuntimeException
+import com.badlogic.gdx.utils.SerializationException
 import com.unciv.Constants
 import com.unciv.UncivGame
+import com.unciv.logic.UncivShowableException
+import com.unciv.logic.github.Github
+import com.unciv.logic.github.Github.folderNameToRepoName
 import com.unciv.models.translations.tr
 import com.unciv.ui.components.UncivTooltip.Companion.addTooltip
 import com.unciv.ui.components.extensions.UncivDateFormat.formatDate
@@ -23,8 +30,13 @@ import com.unciv.ui.components.input.onDoubleClick
 import com.unciv.ui.popups.ConfirmPopup
 import com.unciv.ui.screens.pickerscreens.PickerScreen
 import com.unciv.utils.Concurrency
+import com.unciv.utils.Log
 import com.unciv.utils.launchOnGLThread
+import java.io.FileNotFoundException
+import java.nio.file.attribute.DosFileAttributes
 import java.util.Date
+import kotlin.io.path.Path
+import kotlin.io.path.readAttributes
 
 
 abstract class LoadOrSaveScreen(
@@ -41,6 +53,11 @@ abstract class LoadOrSaveScreen(
     protected val rightSideTable = Table()
     protected val deleteSaveButton = "Delete save".toTextButton(skin.get("negative", TextButton.TextButtonStyle::class.java))
     protected val showAutosavesCheckbox = CheckBox("Show autosaves".tr(), skin)
+    protected val errorLabel = "".toLabel(Color.RED, alignment = Align.center)
+
+    companion object : Helpers {
+        internal const val saveToClipboardErrorMessage = "Could not save game to clipboard!"
+    }
 
     init {
         savesScrollPane.onChange(::selectExistingSave)
@@ -103,6 +120,7 @@ abstract class LoadOrSaveScreen(
     }
 
     private fun selectExistingSave(saveGameFile: FileHandle) {
+        errorLabel.isVisible = false
         deleteSaveButton.enable()
 
         selectedSave = saveGameFile
@@ -133,6 +151,85 @@ abstract class LoadOrSaveScreen(
             launchOnGLThread {
                 descriptionLabel.setText(textToSet.tr())
             }
+        }
+    }
+
+    /** Show appropriate message for an exception and log the severe cases
+     *  @return isUserFixable */
+    protected fun handleException(ex: Exception, primaryText: String, file: FileHandle? = null): Boolean {
+        val (errorText, isUserFixable) = getLoadExceptionMessage(ex, primaryText, file)
+        if (!isUserFixable)
+            Log.error(primaryText, ex)
+        errorLabel.setText(errorText)
+        errorLabel.isVisible = true
+        return isUserFixable
+    }
+
+    interface Helpers {
+        /** Gets a translated exception message to show to the user.
+         * @param file Optional, used for detection of local read-only files, only relevant for write operations on Windows desktops.
+         * @return The first returned value is the message, the second is signifying if the user can likely fix this problem. */
+        fun getLoadExceptionMessage(ex: Throwable, primaryText: String = "Could not load game!", file: FileHandle? = null): Pair<String, Boolean> {
+            val errorText = StringBuilder(primaryText.tr())
+            errorText.appendLine()
+            var cause = ex
+            while (cause.cause != null && cause is GdxRuntimeException) cause = cause.cause!!
+
+            fun FileHandle.isReadOnly(): Boolean {
+                try {
+                    val attr = Path(file().absolutePath).readAttributes<DosFileAttributes>()
+                    return attr.isReadOnly
+                } catch (_: Throwable) { return false }
+            }
+
+            val isUserFixable = when (cause) {
+                is UncivShowableException -> {
+                    errorText.append(ex.localizedMessage)
+                    true
+                }
+                is SerializationException -> {
+                    errorText.append("The file data seems to be corrupted.".tr())
+                    false
+                }
+                is FileNotFoundException -> {
+                    // This is thrown both for chmod/ACL denials and for writing to a file with the read-only attribute set
+                    // On Windows, `message` is already (and illegally) partially localized.
+                    val localizedMessage = UncivGame.Current.getSystemErrorMessage(5)
+                    val isPermissionDenied = cause.message?.run {
+                        contains("Permission denied") || (localizedMessage != null && contains(localizedMessage))
+                    } == true
+                    if (isPermissionDenied) {
+                        if (file != null && file.isReadOnly())
+                            errorText.append("The file is marked read-only.".tr())
+                        else
+                            errorText.append("You do not have sufficient permissions to access the file.".tr())
+                    }
+                    isPermissionDenied
+                }
+                else -> {
+                    errorText.append("Unhandled problem, [${ex::class.simpleName} ${ex.stackTraceToString()}]".tr())
+                    false
+                }
+            }
+            return Pair(errorText.toString(), isUserFixable)
+        }
+
+        fun loadMissingMods(missingMods: Iterable<String>, onModDownloaded:(String)->Unit, onCompleted:()->Unit) {
+            for (rawName in missingMods) {
+                val modName = rawName.folderNameToRepoName().lowercase()
+                val repos = Github.tryGetGithubReposWithTopic(10, 1, modName)
+                    ?: throw UncivShowableException("Could not download mod list.")
+                val repo = repos.items.firstOrNull { it.name.lowercase() == modName }
+                    ?: throw UncivShowableException("Could not find a mod named \"[$modName]\".")
+                val modFolder = Github.downloadAndExtract(
+                    repo,
+                    UncivGame.Current.files.getModsFolder()
+                )
+                    ?: throw Exception("Unexpected 404 error") // downloadAndExtract returns null for 404 errors and the like -> display something!
+                Github.rewriteModOptions(repo, modFolder)
+                onModDownloaded(repo.name)
+            }
+            onCompleted()
         }
     }
 }
