+++ core/src/com/unciv/ui/screens/savescreens/SaveGameScreen.kt
@@ -2,7 +2,6 @@ package com.unciv.ui.screens.savescreens
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.files.FileHandle
-import com.badlogic.gdx.graphics.Color
 import com.badlogic.gdx.scenes.scene2d.ui.Table
 import com.unciv.UncivGame
 import com.unciv.logic.GameInfo
@@ -22,23 +21,25 @@ import com.unciv.ui.components.input.onActivation
 import com.unciv.ui.components.input.onClick
 import com.unciv.ui.popups.ConfirmPopup
 import com.unciv.ui.popups.ToastPopup
-import com.unciv.ui.screens.savescreens.LoadGameScreen.Companion.getLoadExceptionMessage
 import com.unciv.utils.Concurrency
 import com.unciv.utils.Log
 import com.unciv.utils.launchOnGLThread
 
 
-class SaveGameScreen(val gameInfo: GameInfo) : LoadOrSaveScreen("Current saves") {
-    companion object {
+class SaveGameScreen(private val gameInfo: GameInfo) : LoadOrSaveScreen("Current saves") {
+    private val gameNameTextField = UncivTextField(nameFieldLabelText)
+
+    companion object : Helpers {
         const val nameFieldLabelText = "Saved game name"
         const val saveButtonText = "Save game"
         const val savingText = "Saving..."
         const val saveToCustomText = "Save to custom location"
     }
 
-    private val gameNameTextField = UncivTextField(nameFieldLabelText)
-
     init {
+        errorLabel.isVisible = false
+        errorLabel.wrap = true
+
         setDefaultCloseAction()
 
         rightSideTable.initRightSideTable()
@@ -57,13 +58,15 @@ class SaveGameScreen(val gameInfo: GameInfo) : LoadOrSaveScreen("Current saves")
         addGameNameField()
 
         val copyJsonButton = "Copy to clipboard".toTextButton()
-        copyJsonButton.onActivation { copyToClipboardHandler() }
+        copyJsonButton.onActivation(::copyToClipboardHandler)
         val ctrlC = KeyCharAndCode.ctrl('c')
         copyJsonButton.keyShortcuts.add(ctrlC)
         copyJsonButton.addTooltip(ctrlC)
         add(copyJsonButton).row()
 
         addSaveToCustomLocation()
+        add(errorLabel).width(stage.width / 2).center().row()
+        row() // For uniformity with LoadScreen which has a load missing mods button here
         add(deleteSaveButton).row()
         add(showAutosavesCheckbox).row()
     }
@@ -88,10 +91,13 @@ class SaveGameScreen(val gameInfo: GameInfo) : LoadOrSaveScreen("Current saves")
             // the Gzip rarely leads to ANRs
             try {
                 Gdx.app.clipboard.contents = UncivFiles.gameInfoToString(gameInfo, forceZip = true)
-            } catch (ex: Throwable) {
-                ex.printStackTrace()
                 launchOnGLThread {
-                    ToastPopup("Could not save game to clipboard!", this@SaveGameScreen)
+                    ToastPopup("Current game copied to clipboard!", this@SaveGameScreen)
+                }
+            } catch (ex: Throwable) {
+                Log.error(saveToClipboardErrorMessage, ex)
+                launchOnGLThread {
+                    ToastPopup(saveToClipboardErrorMessage, this@SaveGameScreen)
                 }
             }
         }
@@ -99,11 +105,10 @@ class SaveGameScreen(val gameInfo: GameInfo) : LoadOrSaveScreen("Current saves")
 
     private fun Table.addSaveToCustomLocation() {
         val saveToCustomLocation = saveToCustomText.toTextButton()
-        val errorLabel = "".toLabel(Color.RED)
         saveToCustomLocation.onClick {
-            errorLabel.setText("")
             saveToCustomLocation.setText(savingText.tr())
             saveToCustomLocation.disable()
+            errorLabel.isVisible = false
             Concurrency.runOnNonDaemonThreadPool(saveToCustomText) {
 
                 game.files.saveGameToCustomLocation(gameInfo, gameNameTextField.text,
@@ -112,8 +117,7 @@ class SaveGameScreen(val gameInfo: GameInfo) : LoadOrSaveScreen("Current saves")
                     },
                     {
                         if (it !is PlatformSaverLoader.Cancelled) {
-                            errorLabel.setText("Could not save game to custom location!".tr())
-                            it.printStackTrace()
+                            handleException(it, "Could not save game to custom location!")
                         }
                         saveToCustomLocation.setText(saveToCustomText.tr())
                         saveToCustomLocation.enable()
@@ -122,18 +126,16 @@ class SaveGameScreen(val gameInfo: GameInfo) : LoadOrSaveScreen("Current saves")
             }
         }
         add(saveToCustomLocation).row()
-        add(errorLabel).row()
     }
 
     private fun saveGame() {
         rightSideButton.setText(savingText.tr())
+        errorLabel.isVisible = false
         Concurrency.runOnNonDaemonThreadPool("SaveGame") {
             game.files.saveGame(gameInfo, gameNameTextField.text) {
                 launchOnGLThread {
                     if (it != null) {
-                        Log.error("Error saving loading game", it)
-                        val (errorText, _) = getLoadExceptionMessage(it, "Could not save game!", game.files.getSave(gameNameTextField.text))
-                        ToastPopup(errorText, this@SaveGameScreen)
+                        handleException(it, "Could not save game!", game.files.getSave(gameNameTextField.text))
                         rightSideButton.setText(saveButtonText.tr())
                     }
                     else UncivGame.Current.popScreen()
